requirements.txt
django==4.2.2
djangorestframework==3.14.0
django-cors-headers==4.1.0